kafka_pubsub/pub/requirements.txt
imdbpy
kafka-python